+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 0.5.30 (Pending):
+* If repository debug output is enabled in application then print out a helper message
+  when the GwtDataLoaderService is defined. Helps developers remember this feature is
+  present.
 * Associate a symbolic key with each session context.
 
 ## 0.5.29: